refactor(@angular-devkit/build-angular): move bundler context setup into separate file
To reduce the amount of code within the main `application` builder execution function,
the bundler context setup has been moved into a separate file. This also reduces the
amount of imports within the main execution function's module.
